Moves all FFI related to major version 5
