--- build.sh
@@ -1,3 +1,0 @@
-#!/bin/bash 
-cargo build --release  
-go build .
