--- example/go.mod
@@ -1,7 +1,0 @@
-module example-govrl
-
-go 1.18
-
-replace github.com/gh123man/go-vrl => ../
-
-require github.com/gh123man/go-vrl v0.0.0-00010101000000-000000000000
