--- go.mod
@@ -1,3 +1,0 @@
-module github.com/gh123man/go-vrl
-
-go 1.19
