+++ v5/example/go.mod
@@ -1,0 +1,7 @@
+module example-govrl
+
+go 1.18
+
+replace github.com/gh123man/go-vrl/v5 => ../v5
+
+require github.com/gh123man/go-vrl/v5 v5.0.0
