+++ v5/example/main.go
@@ -4,7 +4,7 @@ import (
 	"fmt"
 	"log"
 
-	govrl "github.com/gh123man/go-vrl"
+	govrl "github.com/gh123man/go-vrl/v5"
 )
 
 func main() {
