+++ v5/go.mod
@@ -1,0 +1,3 @@
+module github.com/gh123man/go-vrl/v5
+
+go 1.19
